pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "AI Agent Project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
